fix: lint of gauge name validation
